requirements.txt
jinja2
GitPython
requests
PyYAML
docopt
